+++ jeap-error-handling-service/src/main/java/ch/admin/bit/jeap/errorhandling/infrastructure/persistence/ErrorCountByClusterNameResult.java
@@ -1,0 +1,4 @@
+package ch.admin.bit.jeap.errorhandling.infrastructure.persistence;
+
+public record ErrorCountByClusterNameResult(String clusterName, Long errorCount) {
+}