+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [14.1.0] - 2025-11-03
+
+### Changed
+
+- Update parent from 27.4.0 to 27.4.1
+- Consumer only active on default cluster
+- Added new metric eh_errors_by_cluster
+
 ## [14.0.0] - 2025-10-30
 
 ### Changed